Fix global-overflow found in HW ASAN.
1) DL Get Version response info length is 15 bytes, but trying to copy in
buffer size 11 bytes. Hence Buffer size changed to 15 bytes.
2) set_cfg_cmd_len can be maximum 0xFF + header byte size. But data type is
set to uint8_t, which can hold only 0xFF length. Hence data type changed to
uint16_t

+++ SN100x/halimpl/hal/phNxpNciHal_ext.cc
@@ -1118,7 +1118,7 @@ NFCSTATUS request_EEPROM(phNxpNci_EEPROM_info_t* mEEPROM_info) {
   uint8_t cur_value = 0, len = 5;
   uint8_t b_position = 0;
   bool_t update_req = false;
-  uint8_t set_cfg_cmd_len = 0;
+  uint16_t set_cfg_cmd_len = 0;
   uint8_t* set_cfg_eeprom, *base_addr;
 
   mEEPROM_info->update_mode = BITWISE;
